third_party/blink/web_tests/platform/linux-chrome/external/wpt/serial/requestPort/sandboxed_iframe.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] RequestPort from a sandboxed iframe is valid.
  assert_equals: expected (object) object "[object Object]" but got (string) "FAIL: NotFoundError: Failed to execute 'requestPort' on 'Serial': No port selected by the user."
Harness: the test ran to completion.
